feat(android): Restore TopAppBar "Last Contact" pill (2.9.2)
Move the device-heartbeat indicator from the full-width "Device" section
2.9.1 placed at the top of Home and History back into a small rounded
pill in the TopAppBar — denser surface, ambient at-a-glance from every
main tab. Same staleness logic and same data plumbing
(LiveClock-driven `DeviceCheckIn.format`, app-scoped staleness check,
11-min threshold) — only the rendering location and treatment change.

Pill uses neutral M3 tokens (`surfaceVariant` fresh, `errorContainer`
stale) instead of the door-state tint the original pre-2.7/2.8 pill had,
to fit the M3 chrome that arrived in the 2.7/2.8 redesign. Hidden on
the Function list and Diagnostics sub-screens.

The "Not receiving updates from server" alert with the Retry button
stays on Home and History — the pill is the ambient glance, the alert
is the actionable banner. Different concerns; complementary, not
redundant.

Net code change: -371 lines (the section was heavier than the pill).

Files:
- New `ui/TitleBarCheckInPill.kt` (renderer + 4 previews) and
  `ui/home/DeviceCheckIn.kt` (formatter + display data, extracted
  from the deleted `DeviceCheckInSection.kt` so existing 14
  `DeviceCheckInTest` tests keep passing without modification).
- `ui/Main.kt` collects `DoorViewModel.currentDoorEvent` +
  `nowEpochSeconds`, formats once, renders pill in `TopAppBar.actions`.
- `ui/HomeContent.kt`, `ui/DoorHistoryContent.kt` (bridges) and
  `ui/home/HomeContent.kt`, `ui/history/HistoryContent.kt` (stateless)
  drop the `deviceCheckIn` parameter and the in-content "Device" item.
- `ui/OldLastCheckInBanner.kt` deleted; the History stale banner with
  Retry is now an inline `ErrorCard(...)` in `DoorHistoryContent.kt`.
- `version.properties` 2.9.1 → 2.9.2; matching CHANGELOG entry.
