+++ AndroidGarage/androidApp/src/main/java/com/chriscartland/garage/ui/Main.kt
@@ -22,8 +22,10 @@ import androidx.compose.animation.core.tween
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.togetherWith
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.filled.DateRange
@@ -38,6 +40,7 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.remember
@@ -48,6 +51,7 @@ import androidx.lifecycle.viewmodel.navigation3.rememberViewModelStoreNavEntryDe
 import androidx.navigation3.runtime.entryProvider
 import androidx.navigation3.runtime.rememberSaveableStateHolderNavEntryDecorator
 import androidx.navigation3.ui.NavDisplay
+import com.chriscartland.garage.ui.home.DeviceCheckIn
 import com.chriscartland.garage.ui.settings.DiagnosticsScreen
 import com.chriscartland.garage.ui.theme.AppTheme
 import com.chriscartland.garage.usecase.AppLoggerViewModel
@@ -124,6 +128,13 @@ fun AppNavigation(
     // For tab navigation this is fine — process death just restarts on Home tab.
     val backStack = remember { mutableStateListOf<Screen>(Screen.Home) }
 
+    val currentDoorEvent by doorViewModel.currentDoorEvent.collectAsState()
+    val nowEpochSeconds by doorViewModel.nowEpochSeconds.collectAsState()
+    val deviceCheckIn = DeviceCheckIn.format(
+        lastCheckInSeconds = currentDoorEvent.data?.lastCheckInTimeSeconds,
+        nowSeconds = nowEpochSeconds,
+    )
+
     Scaffold(
         topBar = {
             val currentScreen = backStack.lastOrNull()
@@ -151,11 +162,12 @@ fun AppNavigation(
                         }
                     }
                 },
-                // Device-heartbeat indicator was retired from the TopAppBar in
-                // PR C of the FCM/clock/heartbeat series; the same data is now
-                // surfaced as a "Device" section row inside Home and History
-                // (see `DeviceCheckInSection`), in the new M3 sectioned-list
-                // visual language. The TopAppBar carries title + nav only.
+                actions = {
+                    if (!isSubScreen) {
+                        TitleBarCheckInPill(display = deviceCheckIn)
+                        Spacer(modifier = Modifier.width(8.dp))
+                    }
+                },
             )
         },
         bottomBar = {
