+++ AndroidGarage/androidApp/src/main/java/com/chriscartland/garage/ui/TitleBarCheckInPill.kt
@@ -1,0 +1,169 @@
+/*
+ * Copyright 2024 Chris Cartland. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package com.chriscartland.garage.ui
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.LocalContentColor
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.scale
+import androidx.compose.ui.graphics.ColorFilter
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.chriscartland.garage.R
+import com.chriscartland.garage.ui.home.DeviceCheckInDisplay
+import com.chriscartland.garage.ui.theme.PreviewSurface
+
+/**
+ * Compact rounded-pill device-heartbeat indicator that lives in the
+ * TopAppBar's `actions` slot. Antenna icon + duration text when fresh;
+ * disconnected icon + error coloring when stale (>11 min).
+ *
+ * Restored in 2.9.2 in place of the full-width "Device" section that the
+ * 2.9.1 redesign placed on Home and History — same staleness logic
+ * (`DeviceCheckIn.format`), denser surface, ambient at-a-glance from every
+ * main tab.
+ *
+ * Stateless: pass a pre-formatted [DeviceCheckInDisplay]. The duration label
+ * and `isStale` boolean both come from the LiveClock-driven flow in the
+ * caller (typically `Main.kt`).
+ *
+ * Uses neutral M3 tokens: `surfaceVariant` + `onSurfaceVariant` when fresh,
+ * `errorContainer` + `onErrorContainer` when stale.
+ */
+@Composable
+fun TitleBarCheckInPill(
+    display: DeviceCheckInDisplay,
+    modifier: Modifier = Modifier,
+) {
+    val pillShape = RoundedCornerShape(50)
+    val backgroundColor = if (display.isStale) {
+        MaterialTheme.colorScheme.errorContainer
+    } else {
+        MaterialTheme.colorScheme.surfaceVariant
+    }
+    val contentColor = if (display.isStale) {
+        MaterialTheme.colorScheme.onErrorContainer
+    } else {
+        MaterialTheme.colorScheme.onSurfaceVariant
+    }
+    val showText = display.durationLabel != "No data yet"
+    CompositionLocalProvider(LocalContentColor provides contentColor) {
+        Box(
+            modifier = modifier
+                .background(color = backgroundColor, shape = pillShape)
+                .padding(start = 8.dp, end = 4.dp),
+        ) {
+            Row(verticalAlignment = Alignment.CenterVertically) {
+                if (showText) {
+                    Text(
+                        text = display.durationLabel,
+                        style = MaterialTheme.typography.labelSmall,
+                    )
+                    Spacer(modifier = Modifier.width(4.dp))
+                }
+                Image(
+                    modifier = Modifier.scale(0.7f),
+                    painter = if (display.isStale) {
+                        painterResource(id = R.drawable.outline_signal_disconnected_24)
+                    } else {
+                        painterResource(id = R.drawable.baseline_cell_tower_24)
+                    },
+                    colorFilter = ColorFilter.tint(LocalContentColor.current),
+                    contentDescription = if (display.isStale) {
+                        "Device offline"
+                    } else {
+                        "Device online"
+                    },
+                )
+            }
+        }
+    }
+}
+
+@Preview
+@Composable
+fun TitleBarCheckInPillFreshPreview() {
+    PreviewSurface {
+        Row(verticalAlignment = Alignment.CenterVertically) {
+            TitleBarCheckInPill(
+                display = DeviceCheckInDisplay(
+                    durationLabel = "30 sec ago",
+                    isStale = false,
+                ),
+            )
+        }
+    }
+}
+
+@Preview
+@Composable
+fun TitleBarCheckInPillAgingPreview() {
+    PreviewSurface {
+        Row(verticalAlignment = Alignment.CenterVertically) {
+            TitleBarCheckInPill(
+                display = DeviceCheckInDisplay(
+                    durationLabel = "5 min 20 sec ago",
+                    isStale = false,
+                ),
+            )
+        }
+    }
+}
+
+@Preview
+@Composable
+fun TitleBarCheckInPillStalePreview() {
+    PreviewSurface {
+        Row(verticalAlignment = Alignment.CenterVertically) {
+            TitleBarCheckInPill(
+                display = DeviceCheckInDisplay(
+                    durationLabel = "23 min ago",
+                    isStale = true,
+                ),
+            )
+        }
+    }
+}
+
+@Preview
+@Composable
+fun TitleBarCheckInPillNoDataPreview() {
+    PreviewSurface {
+        Row(verticalAlignment = Alignment.CenterVertically) {
+            TitleBarCheckInPill(
+                display = DeviceCheckInDisplay(
+                    durationLabel = "No data yet",
+                    isStale = false,
+                ),
+            )
+        }
+    }
+}
