+++ AndroidGarage/androidApp/src/main/java/com/chriscartland/garage/ui/home/HomeContent.kt
@@ -151,7 +151,6 @@ fun HomeContent(
     modifier: Modifier = Modifier,
     remoteButtonState: RemoteButtonState = RemoteButtonState.Ready,
     alerts: List<HomeAlert> = emptyList(),
-    deviceCheckIn: DeviceCheckInDisplay? = null,
     isRefreshing: Boolean = false,
     onRefresh: () -> Unit = {},
     onAlertAction: (HomeAlert) -> Unit = {},
@@ -184,14 +183,6 @@ fun HomeContent(
             item(key = "status") {
                 HomeSection(label = "Status") {
                     HomeStatusCardBody(status = status)
-                }
-            }
-
-            if (deviceCheckIn != null) {
-                item(key = "device") {
-                    HomeSection(label = "Device") {
-                        DeviceCheckInSection(display = deviceCheckIn)
-                    }
                 }
             }
